requirements.txt
bcrypt==5.0.0
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
Faker==37.5.3
fonttools==4.62.0
kiwisolver==1.5.0
matplotlib==3.10.8
mysql-connector-python==9.4.0
neat-python==1.1.0
numpy==2.3.2
packaging==26.0
pandas==2.3.1
pillow==12.1.1
pygame==2.6.1
pymongo==4.14.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
six==1.17.0
tzdata==2025.2
